move helper functions to helpers file
